requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
mlxtend
pyyaml
papermill
streamlit